pyproject.toml
[project]
name = "palona-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arize-phoenix-otel>=0.15.0",
    "clip>=0.2.0",
    "datasets>=4.8.4",
    "faiss-cpu>=1.13.2",
    "flask>=3.1.3",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-experimental>=0.4.1",
    "langchain-huggingface>=1.2.1",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-openai>=1.1.12",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.3",
    "markdown>=3.10.2",
    "mcp>=1.26.0",
    "nltk>=3.9.4",
    "numpy>=2.4.3",
    "open-clip-torch>=3.3.0",
    "openai>=2.30.0",
    "openinference-instrumentation-openai>=0.1.43",
    "pandas>=3.0.1",
    "pgvector>=0.4.2",
    "psycopg2>=2.9.11",
    "pydantic>=2.12.5",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.3.0",
    "sqlalchemy>=2.0.48",
    "torch>=2.11.0",
    "transformers>=5.4.0",
]
